Everything works, except the scroll
The list is reffered to in the right way now and data isn't corrupted as before.

The "listeners", or what you can call them have been made a little simpler.
The updateCursor() in Display.h has got an update in combination with the scroll functions. Even though they are closer than before, there is still a bug. Scroll works but top and bottom txt does not remain true all the time when scrolling. They get confused with oneanother, but the order is still right. Probably a 1 and 0 or small part in those which are at fault.

The timer also works for the Serial.print()

Notes have been added here and there where needed